nova-core: factor out firmware file path construction
This removes duplicated code for creating the file system pathname
for firmware files. The pathname is intricate and easy to get wrong,
so a single source of truth is valuable here.

There is a tiny run time cost (in a non-hot-path part of the code) to
doing so, which is: we must use .push() methods, instead of a  format
string operation. This is because one call site is evaluated at compile
time, while the other is at run time.

Signed-off-by: John Hubbard <jhubbard@nvidia.com>