[RFC] nova-core: create debugfs root in PCI init closure
Create the 'nova_core' root debugfs entry when the driver loads.

The directory is not removed when the driver unloads and so
reloading the driver doesn't work.

Normally, non-const global variables need to be protected by a
mutix.  Instead, we use unsafe code, as we know the entry is never
modified after the driver is loaded.

Dropping the Mutex solves the lifetime issue of the Mutex guard, which
would otherwise have required the use of `pin_init_scope`.

Signed-off-by: Timur Tabi <ttabi@nvidia.com>