[RFC] rust: module: add init closure to module_pci_driver macro
Mostly created by Cursor.  The closure is executed when the driver is
loaded.  We probably need another one for when the driver is unloaded.

This is used for PCI drivers that perform custom operations when the
driver is loaded.  Frankly, I'm not a fan of this approach, I would
rather such drivers implement a standard module init function and then
manually register the PCI device.  However, the trend appears to be
to create macros for everything, no matter how complicated they get,
and then just make them more complicated to add flexibility.  So here
we are.

Signed-off-by: Timur Tabi <ttabi@nvidia.com>