rust: debugfs: add lookup contructor
Add a new constructor for Dir that uses the debugfs_lookup() API to
obtain a reference to an existing debugfs directory entry.

The key difference from Dir::new() and Dir::subdir() is the cleanup
semantics: when a Dir obtained via lookup() is dropped, it calls
dput() to release the reference rather than debugfs_remove() which
would delete the directory.

To implement this cleanup distinction, the Entry class now includes
an is_lookup boolean that specifies how the entry was created and
therefore how it should be dropped.

Signed-off-by: Timur Tabi <ttabi@nvidia.com>